Move Goto actions to Parallel/Actions
